raise HTTPError instead of calling write_final()
Prior to this commit, there was a method called write_final() in the
BaseHandler class. This method would write back to the client, and
finish with self.finish(). However, this meant that status codes weren't
returned to the client.

A cleaner way to implement this is to raise tornado.web.HTTPError with a
HTTP status code and log the error message. This commit does that.