modify redis use to rely more on logstash
Prior to this commit, we were storing the webhook payload and UUIDs in
Redis permanently. Elasticsearch is much better suited for this task.

This commit modifies the data spec for Redis to ship stringified JSON to
a 'hookshot' key so that it can be consumed by Logstash. It continues to
maintain lists of UUIDs associated with repos, and a 'recent' list, for
reporting purposes.